requirements.txt
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
mixins==0.1.4
PyMySQL==1.0.2
pytz==2022.1
serializers==0.2.4
sqlparse==0.4.2
uwsgi
